LayoutTests/imported/w3c/web-platform-tests/accname/name/comp_name_from_content.tentative-expected.txt

heading with link referencing image using aria-labelledby, that in turn references text element via aria-labelledby

link

heading with link referencing image using aria-labelledby, that in turn references itself and another element via aria-labelledby

link


PASS heading with link referencing image using aria-labelledby, that in turn references text element via aria-labelledby
FAIL heading with link referencing image using aria-labelledby, that in turn references itself and another element via aria-labelledby assert_equals: <h3 data-expectedlabel="image link" data-testname="heading with link referencing image using aria-labelledby, that in turn references itself and another element via aria-labelledby" class="ex">
  <a href="#" aria-labelledby="nested_image_label4">
    <span class="note" id="crossref_link2">link</span><!-- this text is skipped the first time around because of aria-labelledby on parent element -->
  </a>
  <!-- but it's picked up again (after the self-referencial image alt) in inverse order b/c of cross-referencial aria-labelledby edge case -->
  <img id="nested_image_label4" alt="image" aria-labelledby="nested_image_label4 crossref_link2" src="data:image/gif;base64,R0lGODlhAQABAIAAAP///wAAACH5BAEAAAAALAAAAAABAAEAAAICRAEAOw==">
</h3> expected "image link" but got "image image link"
